+++ lib/icon_to_text_extension_codespark.dart
@@ -1,3 +1,4 @@
 library;
 
 export 'src/icon_text_extension.dart';
+export 'src/icon_text_label.dart';